Style recent card
