+++ frontend/scripts/react-components/explore/featured-cards/featured-cards.component.jsx
@@ -5,68 +5,10 @@ import PropTypes from 'prop-types';
 import capitalize from 'lodash/capitalize';
 import { useTransition, animated } from 'react-spring/web.cjs';
 import ResizeListener from 'react-components/shared/resize-listener.component';
-import { ImgBackground } from 'react-components/shared/img';
 import cx from 'classnames';
+import FeaturedCard from './featured-card.component';
 
 import 'react-components/explore/featured-cards/featured-cards.scss';
-
-const FeaturedCard = ({ card, openModal, step }) => {
-  const { title, subtitle, countryName, commodityName } = card;
-  const stepIds = { 1: '', 2: `-${commodityName}`, 3: `-${commodityName}-${countryName}` };
-  const cardStepId = stepIds[step];
-  return (
-    <div className="c-featured-card">
-      <ImgBackground
-        as="button"
-        onClick={() => openModal(card)}
-        className="featured-card-button"
-        data-test={`featured-card${cardStepId}`}
-        src={`'/images/featured-links/${countryName}.svg'`}
-      >
-        <Text
-          variant="mono"
-          align="center"
-          weight="bold"
-          size="lg"
-          transform="uppercase"
-          color="white"
-        >
-          <span className="featured-card-country-name">{countryName}</span>
-          <span className="featured-card-commodity-name">{commodityName}</span>
-        </Text>
-        <Text
-          variant="mono"
-          align="center"
-          weight="bold"
-          transform="uppercase"
-          color="white"
-          lineHeight="lg"
-          className="featured-card-text"
-          title={title}
-        >
-          {title}
-        </Text>
-        <Text
-          variant="mono"
-          align="center"
-          transform="uppercase"
-          color="white"
-          weight="bold"
-          className="featured-card-text"
-          title={subtitle}
-        >
-          {subtitle}
-        </Text>
-      </ImgBackground>
-    </div>
-  );
-};
-
-FeaturedCard.propTypes = {
-  step: PropTypes.number,
-  card: PropTypes.object.isRequired,
-  openModal: PropTypes.func.isRequired
-};
 
 const FeaturedCards = props => {
   const { countryName, commodityName, step, cards, openModal } = props;
