+++ frontend/scripts/react-components/explore/featured-cards/featured-cards.scss
@@ -53,6 +53,10 @@ $card-min-width: 180px;
   background-color: $strong-pink;
   border-radius: 5px;
 
+  &.recent-card {
+    background-color: $charcoal-grey;
+  }
+
   .featured-card-button {
     height: 100px;
     width: 100%;
@@ -62,6 +66,12 @@ $card-min-width: 180px;
     background-size: auto 90%;
     background-position: center;
     background-repeat: no-repeat;
+  }
+
+  &.recent-card {
+    .featured-card-button {
+      padding: 25px 20px 20px;
+    }
   }
 
   .featured-card-country-name {
@@ -81,9 +91,11 @@ $card-min-width: 180px;
     }
   }
 
-  .featured-card-text {
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
+  &:not(.recent-card) {
+    .featured-card-text {
+      white-space: nowrap;
+      overflow: hidden;
+      text-overflow: ellipsis;
+    }
   }
 }
